% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affyCustomCdf.R
\docType{package}
\name{affyCustomCdf}
\alias{affyCustomCdf}
\alias{affyCustomCdf-package}
\title{affyCustomCdf: A package for creating custom cdfs of Affymetrix GeneChip
arrays based on the selected genome and annotation.}
\description{
The affyCustomCdf package provides one function:
CreateAffyCustomCdf.
}
\section{affyCustomCdf functions}{

The affyCustomCdf functions ...
}

\author{
Ernur Saka
}
