% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affyCustomCdfClassDef.R
\name{affyCustomCdfClass}
\alias{affyCustomCdfClass}
\title{An S4 class to represent the Affymetrix Probe alignmnets to an annotation
 An S4 class is used to create a custom CDF sturucture}
\description{
'Long descrition of the function will be here!!!!!!!!!!!!!!!!!
}
\section{Slots}{

\describe{
\item{\code{total}}{is a data table which keeps the aligment results of probes to the
annotations}

\item{\code{modifiedTable}}{is a data tabe which  contains the remainder of total
after probe shaing procedues applied.}

\item{\code{orginalCdfName}}{the name of the original cdf file.}

\item{\code{probeAlignmentFile}}{name of the file that contains the genome aligment
results of probes.}

\item{\code{gtfFileName}}{gtf file name that is being used as a annotation source.}

\item{\code{newCDFName}}{The name of the created custom CDF}

\item{\code{controlProbeSetNumber}}{The number of control probe set located in the
custom CDF}

\item{\code{probeLength}}{The number of probe bases.}

\item{\code{minProbeSetNumber}}{Desired minimum number of probes in a probe set.}

\item{\code{unitType}}{The type of the probe set units}

\item{\code{SD}}{Direction relation between annotation and probe. Values 0 no
directional relation checked}

\item{\code{type}}{type of the created custom CDF. It can be regionG, gene or
transcript.}

\item{\code{MM}}{It show wheather the original CDF has mismatched probes or not.}
}}

\author{
Ernur Saka
}
