% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createAffyCustomCdf.R
\name{createAffyCustomCdf}
\alias{createAffyCustomCdf}
\title{A function to create Affymetrix GeneChip custom Chip Description File (CDF)
createAffyCustomCdf function creates custom CDF for Affymetrix GeneChips
based on the supplied annotations of interest in a General/Gene Transfer
Format(GTF)}
\arguments{
\item{originalCdfName}{String The original CDF file of the selected
Affymetrix GeneChip technology. It can be obtained from Affymetrix NetAffx
web site. i.e. Rat230_2.cdf file for Affymetrix GeneChip Rat 230}

\item{probeAlignmentFile}{String The tab separated probe alignment file.
The file has five columns, in order: X Location of a Probe, Y location
of a probe, Sense/antisense of a probe (-/+), chromosome name, Start Point.
i.e. 335 337 - 8 22161426. Mapping of probes onto a genome cab be performed
via Bowtie and tab separated file of exact columns, can be produced via R
scripts}

\item{gtfFileName}{String The General/Gene Transfer Format(GTF) file name.}

\item{newCDFName}{String Name of the created custom CDF. Default name is
orginalCdfName_min_minProbeSetNumber_D_SD.cdf}

\item{reportFile}{String Name of the created report file name. Default name
is report_orginalCdfNameorginalCdfName.txt}

\item{controlProbeSetNumber}{Number The number of control probe sets in
the original CDF. Control probe set names usually starts with AFFX prescript.
If the number of control probe set is given, they will be included in the
custom CDF without changes otherwise they will not be included. Default value
is 0 (not included).}

\item{minProbeSetNumber}{number The minimum number of probes in a probe set.
Probe sets with less than the minimum number will not be included in the
custom CDF. Default value is 3.}

\item{probeLength}{Number Number of bases in a probe. This information can be
obtained via checking a sequence of a probe belongs to an Affymetrix
GeneChip. Default value is 25.}

\item{SD}{Number Sense/antisense relationship between probes and annotations.
Possible values are 0 (No direction), 1 (Same direction) and 0 (opposite
direction). When SD is 0, direction will not be considered during
probes to annotations mapping. When SD is 1, mapping is performed between
same directions such as sense probes are being mapped to sense annotations.
When SD is 0, mapping is performed between opposite directions such as sense
probes are being mapped to anti-sense annotations. Default values is 1 (Same
direction).}

\item{type}{String The type of the created custom CDF. Options are regionG,
gene, transcript. When regionG is selected, probe sets are designed to target
a specific region (exon, UTR) of a gene. When gene is selected, probe sets
are designed to target genes. When transcript is selected, probe sets are
designed to target a specific transcript of a gene. Default value is regionG.}

\item{transcriptShare}{Bool It defines whether to allow probe share
between transcripts of a gene or not. It is only used when transcript is
selected as the type. Possible values are FALSE for not allowing probe share
and TRUE for allowing probe share. Default value is FALSE.}

\item{junction}{Bool It defines whether we are adding probes to
probe sets that map onto a junction of a specific region of a gene or not. It
is used when regionG or gene option is selected. Options are FALSE for no
junction and TRUE to allow junctions. Default value is FALSE.}

\item{uniqueProbe}{Bool It defines whether to use probes that map
onto more than one annotations or not. It is an option for user to examine
unique probes to one specific annotation. It can only be used with regionG
option. Possible values are FALSE for not unique, TRUE for unique. Default
value is FALSE.}
}
\value{
If all inputs are proper, createAffyCustomCdf returns TRUE and
produces two output files. One of the files is the ASCII custom CDF file and
the other one is the report file.
}
\description{
createAffyCustomCdf function creates custom CDFs via removing nonspecific
probes, updating probe target mapping based on the supplied genome
information and grouping probes into region (exon, UTR), gene and transcript
levels. User has to supply annotations as a GTF file (Ensemble format), probe
start points in a tab separated text file and the original CDF file of a
specific Affymetrix GeneChip. It also takes several other parameters includes
annotation level (region, gene, transcript), annotation direction, number of
probes per probe set. When default parameters are being used, the produced
custom CDF will consists of region based probe sets with minimum number of
three probes per probe set. Probe sets will not contain original control
probe sets and probes aligned to junctions will not be included in the probe
sets. The number of probes, will be assumed 25 and alignment will be
performed same directed probes and annotations.
}
\examples{
pathLib = system.file(package="affyCustomCdf")
pathLib = file.path(pathLib, "scripts")
cdf = file.path(pathLib, "Rat230_2Sort.cdf")
probeMap = file.path(pathLib, "rat230Short.txt")
gtf = file.path(pathLib, "Rattus_norvegicus.Rnor_6.0.85Sort.gtf")
createAffyCustomCdf(cdf,probeMap,gtf)
}
\author{
Ernur Saka
}
